Linux compatibility
Former-commit-id: f8c3d5557699e961c22be725f44308f50f72a410
Former-commit-id: 4ae8b97bcd29fdeb696035f65464efa2f687973d [formerly 9fda82dec5f9a3086410934fa775ad2d9cb69bf3]
Former-commit-id: e122b327a56e3485c2727acb02ac55cbd8c9cc56

+++ modules/util/source/hip/util/conv.d
@@ -180,7 +180,7 @@ TO to(TO, FROM)(FROM f) pure nothrow
 }
 
 /// This function can be called at compilation time without bringing runtime
-string toString(int x) pure nothrow @safe
+string toString(long x) pure nothrow @safe
 {
     enum numbers = "0123456789";
     bool isNegative = x < 0;
